Fix in Camera::pointUnderPixel introduced by switch to double

+++ QGLViewer/camera.cpp
@@ -948,7 +948,7 @@ void Camera::interpolateTo(const Frame& fr, qreal duration)
  to your scene. Loose boundaries will result in imprecision along the viewing direction. */
 Vec Camera::pointUnderPixel(const QPoint& pixel, bool& found) const
 {
-	qreal depth;
+	float depth;
 	// Qt uses upper corner for its origin while GL uses the lower corner.
 	glReadPixels(pixel.x(), screenHeight()-1-pixel.y(), 1, 1, GL_DEPTH_COMPONENT, GL_FLOAT, &depth);
 	found = depth < 1.0;
